sensor maintanence routine

+++ sensor firmware/gateway_espnow/global_defines.h
@@ -22,10 +22,10 @@
 
 typedef struct  {
   void * next;
-  double fft_values [16];
+  float fft_values [16];
   int fft_range;
   double decibels;
-  long timestamp;
+  unsigned int timestamp;
 } sending_list;
 
 typedef struct  {
