+++ sensor firmware/get_mic_data_and_send_fft/global_defines.h
@@ -23,7 +23,7 @@
 
 typedef struct  {
   void * next;
-  double fft_values [16];
+  float fft_values [16];
   int fft_range;
   double decibels;
   unsigned int timestamp;
